requirements.txt
arxiv
openai
boto3
notion_client
python-dotenv